refactor: fold scope-breadth rank into order-by-scope
Address PR #368 review: a standalone module for the scope-breadth map is
overkill — export SCOPE_BREADTH_RANK from order-by-scope.ts and import it
into the narrower-scope removal rule.
